requirements.txt
tqdm==4.64.0
scipy==1.8.1
scikit-image==0.19.3
PyYAML==6.0
Pillow==9.1.1
pandas==1.4.3
numpy==1.22.4
networkx==2.8.4
matplotlib==3.5.2
libpysal==4.6.2